# ms-order/src/main/resources/swagger.yml
openapi: 3.0.1
info:
  title: MS Order
  description: Microserviço de gerenciamento de pedidos
  version: 1.0.0
servers:
  - url: http://localhost:8081/api/pedidos
paths:
  /pedidos:
    get:
      tags:
        - Order
      summary: Listar pedidos
      operationId: getAllOrderns
      parameters:
        - name: sort
          in: query
          required: false
          schema:
            type: string
        - name: cpf
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDTO'
    post:
      tags:
        - Order
      summary: Cadastrar pedidos
      operationId: createNewOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
  /pedidos/{id}:
    get:
      tags:
        - Order
      summary: Buscar pedidos por ID
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
    delete:
      tags:
        - Order
      summary: Excluir pedidos
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - Order
      summary: Atualizar pedidos
      operationId: updateOrderInformation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
  /item/{id}:
    patch:
      tags:
        - item-controller
      operationId: updateItems
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDTO'
components:
  schemas:
    Address:
      type: object
      properties:
        id:
          type: integer
          format: int64
        street:
          type: string
        number:
          type: string
        district:
          type: string
        city:
          type: string
        state:
          type: string
        cep:
          type: string
    ItemRequest:
      type: object
      properties:
        nome:
          type: string
        valor:
          type: number
        descricao:
          type: string
        data_de_validade:
          type: string
          format: date-time
    OrderRequest:
      required:
        - cpf
        - endereco
        - items
        - total
      type: object
      properties:
        cpf:
          pattern: ^\d{11}$
          type: string
        items:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/ItemRequest'
        total:
          type: number
          format: double
        endereco:
          $ref: '#/components/schemas/Address'
    AddressDTO:
      type: object
      properties:
        cep:
          type: string
        logradouro:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        localidade:
          type: string
        uf:
          type: string
    ItemDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        data_de_criacao:
          type: string
          format: date-time
        data_de_validade:
          type: string
          format: date-time
        valor:
          type: number
        descricao:
          type: string
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cpf:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemDTO'
        total:
          type: number
          format: double
        endereco:
          $ref: '#/components/schemas/AddressDTO'